requirements.txt
python-dotenv
pywin32
requests
certifi
msal
